Create new databases netdata-health, netdata-aclk Move metadata related functions from sqlite_functions.c to sqlite_metadata Move health related functions to sqlite_health.c Move aclk related functions to sqlite_aclk.c Add functions to support database migrations for netdata-health and netdata-aclk Fix issue on database rebuild to correctly apply indices on chart and dimension Generic functions to init databases and apply config per database connection Configure database lookaside allocation slots to improve performance ("look aside buffer slots" and "look aside buffer slot size")
