+++ database/sqlite/sqlite_aclk_alert.h
@@ -3,7 +3,6 @@
 #ifndef NETDATA_SQLITE_ACLK_ALERT_H
 #define NETDATA_SQLITE_ACLK_ALERT_H
 
-extern sqlite3 *db_meta;
 
 #define SEND_REMOVED_AFTER_HEALTH_LOOPS 3
 #define SEND_CHECKPOINT_AFTER_HEALTH_LOOPS 4
