+++ database/sqlite/sqlite_context.c
@@ -45,60 +45,23 @@ int sql_init_context_database(int memory)
 
     info("SQLite database %s initialization", sqlite_database);
 
-    char buf[1024 + 1] = "";
-    const char *list[2] = { buf, NULL };
-
     int target_version = DB_CONTEXT_METADATA_VERSION;
     if (likely(!memory))
         target_version = perform_context_database_migration(db_context_meta, DB_CONTEXT_METADATA_VERSION);
 
-    // https://www.sqlite.org/pragma.html#pragma_auto_vacuum
-    // PRAGMA schema.auto_vacuum = 0 | NONE | 1 | FULL | 2 | INCREMENTAL;
-    snprintfz(buf, 1024, "PRAGMA auto_vacuum=%s;", config_get(CONFIG_SECTION_SQLITE, "auto vacuum", "INCREMENTAL"));
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list)) return 1;
-
-    // https://www.sqlite.org/pragma.html#pragma_synchronous
-    // PRAGMA schema.synchronous = 0 | OFF | 1 | NORMAL | 2 | FULL | 3 | EXTRA;
-    snprintfz(buf, 1024, "PRAGMA synchronous=%s;", config_get(CONFIG_SECTION_SQLITE, "synchronous", "NORMAL"));
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list))  return 1;
-
-    // https://www.sqlite.org/pragma.html#pragma_journal_mode
-    // PRAGMA schema.journal_mode = DELETE | TRUNCATE | PERSIST | MEMORY | WAL | OFF
-    snprintfz(buf, 1024, "PRAGMA journal_mode=%s;", config_get(CONFIG_SECTION_SQLITE, "journal mode", "WAL"));
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list)) return 1;
-
-    // https://www.sqlite.org/pragma.html#pragma_temp_store
-    // PRAGMA temp_store = 0 | DEFAULT | 1 | FILE | 2 | MEMORY;
-    snprintfz(buf, 1024, "PRAGMA temp_store=%s;", config_get(CONFIG_SECTION_SQLITE, "temp store", "MEMORY"));
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list)) return 1;
-    
-    // https://www.sqlite.org/pragma.html#pragma_journal_size_limit
-    // PRAGMA schema.journal_size_limit = N ;
-    snprintfz(buf, 1024, "PRAGMA journal_size_limit=%lld;", config_get_number(CONFIG_SECTION_SQLITE, "journal size limit", 16777216));
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list)) return 1;
-
-    // https://www.sqlite.org/pragma.html#pragma_cache_size
-    // PRAGMA schema.cache_size = pages;
-    // PRAGMA schema.cache_size = -kibibytes;
-    snprintfz(buf, 1024, "PRAGMA cache_size=%lld;", config_get_number(CONFIG_SECTION_SQLITE, "cache size", -2000));
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list)) return 1;
-
-    snprintfz(buf, 1024, "PRAGMA user_version=%d;", target_version);
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list)) return 1;
-
-    if (likely(!memory))
-        snprintfz(buf, 1024, "ATTACH DATABASE \"%s/netdata-meta.db\" as meta;", netdata_configured_cache_dir);
-    else
-        snprintfz(buf, 1024, "ATTACH DATABASE ':memory:' as meta;");
-
-    if(init_database_batch(db_context_meta, DB_CHECK_NONE, 0, list)) return 1;
-
-    if (init_database_batch(db_context_meta, DB_CHECK_NONE, 0, &database_context_config[0]))
+    if (configure_database_params(db_context_meta, target_version))
         return 1;
 
-    if (init_database_batch(db_context_meta, DB_CHECK_NONE, 0, &database_context_cleanup[0]))
+    if (attach_database(db_context_meta, memory ? NULL : "netdata-meta.db", "meta"))
         return 1;
 
+    if (init_database_batch(db_context_meta, &database_context_config[0]))
+        return 1;
+
+    if (init_database_batch(db_context_meta, &database_context_cleanup[0]))
+        return 1;
+
+    info("SQLite database %s initialization completed", sqlite_database);
     return 0;
 }
 
@@ -106,18 +69,10 @@ int sql_init_context_database(int memory)
  * Close the sqlite database
  */
 
-void sql_close_context_database(void)
-{
-    int rc;
-    if (unlikely(!db_context_meta))
-        return;
-
-    info("Closing context SQLite database");
-
-    rc = sqlite3_close_v2(db_context_meta);
-    if (unlikely(rc != SQLITE_OK))
-        error_report("Error %d while closing the context SQLite database, %s", rc, sqlite3_errstr(rc));
-}
+//void sql_close_context_database(void)
+//{
+//    sql_close_database(db_context_meta,"CONTEXT");
+//}
 
 //
 // Fetching data
@@ -451,9 +406,13 @@ int sql_context_cache_stats(int op)
         return 0;
 
     netdata_thread_disable_cancelability();
-    sqlite3_db_status(db_context_meta, op, &count, &dummy, 0);
+    int rc = sqlite3_db_status(db_context_meta, op, &count, &dummy, 0);
     netdata_thread_enable_cancelability();
-    return count;
+
+    if (SQLITE_OK == rc)
+        return count;
+    error_report("METADATA: SQLITE statistics failed with rc = %d, %s", rc, sqlite3_errstr(rc));
+    return 0;
 }
 
 //
@@ -490,7 +449,7 @@ int ctx_unittest(void)
     uuid_t host_uuid;
     uuid_generate(host_uuid);
 
-    initialize_thread_key_pool();
+    init_sqlite_thread_key_pool();
 
     int rc = sql_init_context_database(1);
 
@@ -566,7 +525,11 @@ int ctx_unittest(void)
     ctx_get_context_list(&host_uuid, dict_ctx_get_context_list_cb, NULL);
     info("List context end after delete");
 
-    sql_close_context_database();
+//    sql_close_context_database();
+//    sql_close_database(db_context_meta,"CONTEXT");
+    rc = sqlite3_close_v2(db_context_meta);
+    if (unlikely(rc != SQLITE_OK))
+        error_report("%s: Error while closing the sqlite database: rc %d, error \"%s\"", "CONTEXT", rc, sqlite3_errstr(rc));
 
     return 0;
 }
