+++ database/sqlite/sqlite_db_migration.h
@@ -8,6 +8,8 @@
 
 int perform_database_migration(sqlite3 *database, int target_version);
 int perform_context_database_migration(sqlite3 *database, int target_version);
+int perform_health_database_migration(sqlite3 *database, int target_version);
+int perform_aclk_database_migration(sqlite3 *database, int target_version);
 int table_exists_in_database(const char *table);
 
 #endif //NETDATA_SQLITE_DB_MIGRATION_H
