+++ database/sqlite/sqlite_functions.h
@@ -6,16 +6,6 @@
 #include "daemon/common.h"
 #include "sqlite3.h"
 
-// return a node list
-struct node_instance_list {
-    uuid_t  node_id;
-    uuid_t  host_id;
-    char *hostname;
-    int live;
-    int queryable;
-    int hops;
-};
-
 typedef enum db_check_action_type {
     DB_CHECK_NONE  = 0x0000,
     DB_CHECK_INTEGRITY  = 0x0001,
@@ -27,6 +17,7 @@ typedef enum db_check_action_type {
 #define SQL_MAX_RETRY (100)
 #define SQLITE_INSERT_DELAY (10)        // Insert delay in case of lock
 
+/*
 #define CHECK_SQLITE_CONNECTION(db_meta)                                                                               \
     if (unlikely(!db_meta)) {                                                                                          \
         if (default_rrd_memory_mode != RRD_MEMORY_MODE_DBENGINE) {                                                     \
@@ -35,6 +26,10 @@ typedef enum db_check_action_type {
         error_report("Database has not been initialized");                                                             \
         return 1;                                                                                                      \
     }
+*/
+
+#define SQLITE_LOOKASIDE_SLOTS 192
+#define SQLITE_LOOKASIDE_SLOT_SIZE 8192
 
 SQLITE_API int sqlite3_step_monitored(sqlite3_stmt *stmt);
 SQLITE_API int sqlite3_exec_monitored(
@@ -46,35 +41,26 @@ SQLITE_API int sqlite3_exec_monitored(
     );
 
 // Initialization and shutdown
-int init_database_batch(sqlite3 *database, int rebuild, int init_type, const char *batch[]);
-int sql_init_database(db_check_action_type_t rebuild, int memory);
-void sql_close_database(void);
+int sqlite_library_init(void);
+void sqlite_library_shutdown(void);
+int sqlite_init_databases(db_check_action_type_t mode, bool memory_mode);
+void sqlite_close_databases(void);
+
+int init_database_batch(sqlite3 *database, const char *batch[]);
+int sql_init_metadata_database(db_check_action_type_t rebuild, int memory);
 
 // Helpers
 int bind_text_null(sqlite3_stmt *res, int position, const char *text, bool can_be_null);
 int prepare_statement(sqlite3 *database, const char *query, sqlite3_stmt **statement);
 int execute_insert(sqlite3_stmt *res);
-int exec_statement_with_uuid(const char *sql, uuid_t *uuid);
+int exec_statement_with_uuid(sqlite3 *database, const char *sql, uuid_t(*uuid));
 int db_execute(sqlite3 *database, const char *cmd);
-void initialize_thread_key_pool(void);
-
-// Look up functions
-int get_node_id(uuid_t *host_id, uuid_t *node_id);
-int get_host_id(uuid_t *node_id, uuid_t *host_id);
-struct node_instance_list *get_node_list(void);
-void sql_load_node_id(RRDHOST *host);
-char *get_hostname_by_node_id(char *node_id);
-
-// Help build archived hosts in memory when agent starts
-void sql_build_host_system_info(uuid_t *host_id, struct rrdhost_system_info *system_info);
-DICTIONARY *sql_load_host_labels(uuid_t *host_id);
-
-// TODO: move to metadata
-int update_node_id(uuid_t *host_id, uuid_t *node_id);
+int configure_database_params(sqlite3 *database, int target_version);
+int attach_database(sqlite3 *database, const char *database_path, const char *alias);
+void add_stmt_to_list(sqlite3_stmt *res);
+void init_sqlite_thread_key_pool(void);
+int check_table_integrity(sqlite3 *database, char *table);
 
 void invalidate_node_instances(uuid_t *host_id, uuid_t *claim_id);
-
-// Provide statistics
-int sql_metadata_cache_stats(int op);
 
 #endif //NETDATA_SQLITE_FUNCTIONS_H
